1. sp_cloneTemplate - Types changed int to long 2. Organisation - ShowBleedArea, BleedAreaSize columns added, userdomainkey column removed 3. Marup - Userdomainkey renamed to organisationid

+++ Implementation/WebStoreServices/TemplateService.cs
@@ -285,7 +285,7 @@ namespace MPC.Implementation.WebStoreServices
         }
 
       
-        public int CloneTemplateByTemplateID(int TempID)
+        public long CloneTemplateByTemplateID(long TempID)
         {
             return _templateRepository.CloneTemplateByTemplateID(TempID);
         }
